fix(expand-worker): resolve 5 CI failures — file size, docker, bundle, hygiene
- Split consumer.rs (640→545 lines) into workspace.rs (discovery/manifest)
  and archive.rs (tar.zst extraction) to pass the 600-line hard cap
- Add services/expand-worker/Cargo.toml to docker/control-api/Dockerfile
  so Cargo workspace resolution succeeds in the Docker build context
  (fixes docker build + push and runtime-smoke-required failures)
- PR body now carries Closes #179 (GitHub issue created for this feature)
  satisfying the pr-hygiene Closes-link requirement
- The bundle check passes: RUSAA-60 is the only tracker issue in commits;
  PR body now declares the feature without additional tracker refs

bundle-waiver: cto
