+++ docker/control-api/Dockerfile
@@ -23,6 +23,7 @@ COPY services/ingest-clone/Cargo.toml    services/ingest-clone/Cargo.toml
 COPY services/projector-pg/Cargo.toml    services/projector-pg/Cargo.toml
 COPY services/projector-neo4j/Cargo.toml services/projector-neo4j/Cargo.toml
 COPY services/tombstoner/Cargo.toml      services/tombstoner/Cargo.toml
+COPY services/expand-worker/Cargo.toml   services/expand-worker/Cargo.toml
 
 # Stub sources so `cargo build` caches deps without the full source tree.
 RUN find crates services -name "Cargo.toml" | sed 's|Cargo.toml|src/lib.rs|' | \
